requirements.txt
requests
beautifulsoup4
regex
atexit
python-dotenv